+++ datastore/couchDB/tenantServiceDatastoreCouchDB.go
@@ -998,7 +998,7 @@ func (tsd *TenantServiceDatastoreCouchDB) BulkInsertMonitoredObjects(tenantID st
 		dataProp := genericMO["data"].(map[string]interface{})
 		dataProp["datatype"] = dataType
 		dataProp["createdTimestamp"] = ds.MakeTimestamp()
-		dataProp["lastModifiedTimestamp"] = genericMO["createdTimestamp"]
+		dataProp["lastModifiedTimestamp"] = dataProp["createdTimestamp"]
 
 		data = append(data, genericMO)
 	}
@@ -1060,8 +1060,7 @@ func (tsd *TenantServiceDatastoreCouchDB) BulkUpdateMonitoredObjects(tenantID st
 		dataType := string(tenmod.TenantMonitoredObjectType)
 		dataProp := genericMO["data"].(map[string]interface{})
 		dataProp["datatype"] = dataType
-		dataProp["createdTimestamp"] = ds.MakeTimestamp()
-		dataProp["lastModifiedTimestamp"] = genericMO["createdTimestamp"]
+		dataProp["lastModifiedTimestamp"] = ds.MakeTimestamp()
 
 		data = append(data, genericMO)
 	}